+++ src/Symfony/Bridge/PhpUnit/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * ignore verbosity settings when the build fails because of deprecations
  * added per-group verbosity
+ * added `ExpectDeprecationTrait` to be able to define an expected deprecation from inside a test
 
 5.0.0
 -----